Migrate to Rocket 0.5-rc.1
This was a bit more tricky than expected, mostly due to the database
change. Since database access with Diesel is sync, we need to use the
`rocket_sync_db_pools` crate which requires async handlers. See
https://github.com/SergioBenitez/Rocket/issues/1910 for details.

Besides that, there were a lot of changes in Rocket, but the changes
were mostly mechanical.

Since Rocket now supports stable, we can migrate to a stable toolchain
as well!